media list and paragraph list delete confirmation and other bug resolved

+++ src/Pages/Appeal/Media/List/index.js
@@ -42,11 +42,15 @@ const MediaList = (props) => {
     } else if (error) {
       toast.error(error.message);
       dispatch(clearErrors());
+      window.location.reload();
     }
   }, [loading, error, success]);
 
   const handleDelete = (idx) => {
-    dispatch(deleteAppealMedia(accessToken, idx, id));
+    if (window.confirm("Are you sure you want to delete?")) {
+      dispatch(deleteAppealMedia(accessToken, idx, id));
+    }
+   
   };
 
   if (loading) return <Spinner />;
