+++ src/Pages/Campaigns/Media/List/index.js
@@ -40,11 +40,15 @@ const MediaList = (props) => {
     } else if (error) {
       toast.error(error.message);
       dispatch(clearErrors());
+      window.location.reload();
     }
   }, [loading, error, success]);
 
   const handleDelete = (idx) => {
-    dispatch(deleteCampaignMedia(accessToken, idx, id));
+    if (window.confirm("Are you sure you want to delete?")) {
+      dispatch(deleteCampaignMedia(accessToken, idx, id));
+    }
+   
   };
 
   if (loading) return <Spinner />;
