+++ src/Pages/Projects/Paragraph/List/index.js
@@ -62,11 +62,15 @@ const ParagraphList = (props) => {
     } else if (error) {
       toast.error(error.message);
       dispatch(clearErrors());
+      window.location.reload();
     }
   }, [loading, error, success]);
 
   const handleDelete = (idx) => {
-    dispatch(deleteProjectParagraph(accessToken, idx));
+    if (window.confirm("Are you sure you want to delete?")) {
+      dispatch(deleteProjectParagraph(accessToken, idx));
+    }
+  
   };
 
   const genExtra = (idx) => (
